pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "mqtt2postgres"
dynamic = ["version"]
description = "Minimal MQTT to Postgres ingestor."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "numpy",
  "paho-mqtt",
  "pandas",
  "psycopg2-binary",
  "pyyaml",
  "sqlalchemy",
]

[project.scripts]
mqtt2postgres = "mqtt2postgres.app:main"
mqtt2postgres-twin-config = "mqtt2postgres.twin_config:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
fallback_version = "0.8.0"
version_scheme = "post-release"
local_scheme = "no-local-version"
